# apps/frontend/src/assets/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "企业级 Go+Gin+Gorm Web 后端项目模板",
    "title": "Gin Admin Starter API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "API Support",
      "url": "http://www.swagger.io/support",
      "email": "support@swagger.io"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "paths": {
...
    "/api/v1/rbac/users/{id}/roles/{roleCode}": {
      "delete": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "从指定用户移除指定角色",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["权限管理"],
        "summary": "从用户移除角色",
        "parameters": [
          {
            "type": "integer",
            "description": "用户ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "角色代码",
            "name": "roleCode",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
            }
          }
        }
      }
    },
...
    "/api/v1/users/{id}": {
      "delete": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "删除指定用户",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["用户管理"],
        "summary": "删除用户",
        "parameters": [
          {
            "type": "integer",
            "description": "用户ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
            }
          }
        }
      }
    },
...
    "/roles": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "获取角色列表",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["角色管理"],
        "summary": "获取角色列表",
        "parameters": [
          {
            "type": "integer",
            "default": 1,
            "description": "页码",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "default": 10,
            "description": "每页数量",
            "name": "pageSize",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/go-admin-starter_pkg_response.PageResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/definitions/go-admin-starter_internal_model.Role"
                      }
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "创建角色",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["角色管理"],
        "summary": "创建角色",
        "parameters": [
          {
            "description": "角色信息",
            "name": "role",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/go-admin-starter_internal_model.Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/go-admin-starter_internal_model.Role"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
            }
          }
        }
      }
    },
    "/roles/{id}": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "获取角色",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["角色管理"],
        "summary": "获取角色",
        "parameters": [
          {
            "type": "integer",
            "description": "角色ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/go-admin-starter_internal_model.Role"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
            }
          }
        }
      },
      "put": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "更新角色",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["角色管理"],
        "summary": "更新角色",
        "parameters": [
          {
            "type": "integer",
            "description": "角色ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "角色信息",
            "name": "role",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/go-admin-starter_internal_model.Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/go-admin-starter_internal_model.Role"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "删除角色",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["角色管理"],
        "summary": "删除角色",
        "parameters": [
          {
            "type": "integer",
            "description": "角色ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
            }
          }
        }
      }
    },
    "/roles/{id}/permissions": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "获取角色权限",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["角色管理"],
        "summary": "获取角色权限",
        "parameters": [
          {
            "type": "integer",
            "description": "角色ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/definitions/go-admin-starter_internal_model.Permission"
                      }
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "为角色分配权限",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["角色管理"],
        "summary": "为角色分配权限",
        "parameters": [
          {
            "type": "integer",
            "description": "角色ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "权限信息",
            "name": "req",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
            }
          }
        }
      }
    },
    "/roles/{id}/permissions/{permId}": {
      "delete": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "撤销角色权限",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["角色管理"],
        "summary": "撤销角色权限",
        "parameters": [
          {
            "type": "integer",
            "description": "角色ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "description": "权限ID",
            "name": "permId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/go-admin-starter_pkg_response.Response"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "go-admin-starter_internal_model.Permission": {
      "type": "object",
      "required": ["permCode", "permName", "status"],
      "properties": {
        "children": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/go-admin-starter_internal_model.Permission"
          }
        },
        "createdAt": {
          "type": "string"
        },
        "createdBy": {
          "type": "integer"
        },
        "deletedAt": {
          "$ref": "#/definitions/gorm.DeletedAt"
        },
        "icon": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "menuType": {
          "type": "string"
        },
        "method": {
          "type": "string"
        },
        "parentId": {
          "type": "integer"
        },
        "path": {
          "type": "string"
        },
        "permCode": {
          "type": "string"
        },
        "permName": {
          "type": "string"
        },
        "resourceType": {
          "type": "string"
        },
        "sortOrder": {
          "type": "integer"
        },
        "status": {
          "type": "integer"
        },
        "updatedAt": {
          "type": "string"
        },
        "updatedBy": {
          "type": "integer"
        }
      }
    },
    "go-admin-starter_internal_model.Role": {
      "type": "object",
      "required": ["role_code", "role_name", "status"],
      "properties": {
        "createdAt": {
          "type": "string"
        },
        "createdBy": {
          "type": "integer"
        },
        "deletedAt": {
          "$ref": "#/definitions/gorm.DeletedAt"
        },
        "description": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "role_code": {
          "type": "string"
        },
        "role_name": {
          "type": "string"
        },
        "sort_order": {
          "type": "integer"
        },
        "status": {
          "type": "integer"
        },
        "updatedAt": {
          "type": "string"
        },
        "updatedBy": {
          "type": "integer"
        }
      }
    },
...
    "go-admin-starter_pkg_response.PageResponse": {
      "type": "object",
      "properties": {
        "items": {},
        "page": {
          "type": "integer"
        },
        "pageSize": {
          "type": "integer"
        },
        "total": {
          "type": "integer"
        }
      }
    },
    "go-admin-starter_pkg_response.Response": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer"
        },
        "data": {},
        "message": {
          "type": "string"
        }
      }
    },
    "gorm.DeletedAt": {
      "type": "object",
      "properties": {
        "time": {
          "type": "string"
        },
        "valid": {
          "description": "Valid is true if Time is not NULL",
          "type": "boolean"
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "description": "请输入 JWT Token，格式: Bearer {token}",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}